Move all handlers to their own file
